+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * improved `TestCase::assertLogContains()` method;
 * `deleteLog()` and `getLogFullPath()` methods provided by the `TestCase` class are now deprecated and will be removed
   in a later release;
+* `assertSqlEndsWith()` and `assertSqlEndsNotWith()` methods provided by the `TestCase` class now deprecated and will be
+  removed in a later release. Use instead `assertStringEndsWith()` and `assertStringEndsNotWith()`;
 * `LibraryHelper::analytics()` method was deprecated and has now been removed;
 * suggests `axllent/jquery` and no longer `components/jquery`;
 * `axllent/jquery` added to `MeTools.VendorLinks` configuration;